requirements-codex-test.txt
openai>=1.54.0
openai-agents
pydantic>=2.9.0
python-dotenv>=1.0.1
rich>=13.9.0
typer>=0.12.0
langchain-openai>=0.2.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
requests>=2.31.0
